fixup! TEMP trying to track down these pnode ASSERTs

+++ src/net_processing.cpp
@@ -1020,7 +1020,7 @@ void PeerManagerImpl::ProcessBlockAvailability(NodeId nodeid) {
     CNodeState *state = State(nodeid);
     if (state == nullptr) {
         LogPrintf("%s: ASSERT ERROR. peer=%d\n", __func__, nodeid);
-        fprintf(stderr, "%s: ASSERT ERROR. peer=%ld\n", __func__, nodeid);
+        fprintf(stderr, "%s: ASSERT ERROR. peer=%d\n", __func__, (int)nodeid);
         fflush(stderr);
         return;
     }
@@ -1294,7 +1294,7 @@ void PeerManagerImpl::FinalizeNode(const CNode& node)
         PeerRef peer = RemovePeer(nodeid);
         if (peer == nullptr) {
             LogPrintf("%s: ASSERT ERROR. peer=%d\n", __func__, nodeid);
-            fprintf(stderr, "%s: ASSERT ERROR. peer=%ld\n", __func__, nodeid);
+            fprintf(stderr, "%s: ASSERT ERROR. peer=%d\n", __func__, (int)nodeid);
             fflush(stderr);
             return;
         }
